backend/requirements.txt
# Django & REST
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Database
psycopg2-binary>=2.9,<3.0

# Environment
python-dotenv>=1.0,<2.0

# ML (sklearn Pipelines, DataFrame input)
joblib>=1.3,<2.0
scikit-learn>=1.3,<2.0
pandas>=2.0,<3.0

# Production (optional)
gunicorn>=21.0,<23.0
